Added note for Power BI Report Server
If Power BI Report Server is missing an associated /PowerBI/ urlacl for each name the service will start but users will not be able not be able to upload Power BI reports. The RSPowerBI_ log will report "PBIXSystem.Reflection.TargetInvocationException: Exception has been thrown by the target of an invocation. ---> System.Net.HttpListenerException: Access is denied". When the Service SID is granted the urlacl everything will work fine. 

Would have saved me a lot of time if this was included here :-)